+++ apps/io-node/include/peripherals.h
@@ -10,27 +10,22 @@
 extern "C" {
 #endif
 
+#include "common-peripherals.h"
 #include "stm32f3xx_hal.h"
 
 typedef struct {
+  // Provided by CPU board
+  common_peripherals_t *common_peripherals;
+
   ADC_HandleTypeDef hadc1;
   DMA_HandleTypeDef hdma_adc1;
-  CAN_HandleTypeDef hcan;
   I2C_HandleTypeDef hi2c1;
-  SPI_HandleTypeDef hspi1;
   SPI_HandleTypeDef hspi3;
-  UART_HandleTypeDef huart1;
 } peripherals_t;
 
 peripherals_t *get_peripherals(void);
-void adc1_init(void);
-void can_init(void);
-void i2c1_init(void);
-void spi1_init(void);
-void spi3_init(void);
-void uart1_init(void);
-void dma_init(void);
-void gpio_init(void);
+
+void peripherals_init(void);
 
 #ifdef __cplusplus
 }
