+++ apps/io-node/src/interrupts.c
@@ -7,8 +7,3 @@ void DMA1_Channel1_IRQHandler(void) {
   peripherals_t *peripherals = get_peripherals();
   HAL_DMA_IRQHandler(&peripherals->hdma_adc1);
 }
-
-void USART1_IRQHandler(void) {
-  peripherals_t *peripherals = get_peripherals();
-  HAL_UART_IRQHandler(&peripherals->huart1);
-}
