+++ apps/io-node/src/peripherals.c
@@ -10,7 +10,24 @@
 
 static peripherals_t peripherals;
 
+void adc1_init(void);
+void i2c1_init(void);
+void spi3_init(void);
+void dma_init(void);
+void gpio_init(void);
+
 peripherals_t* get_peripherals(void) { return &peripherals; }
+
+void peripherals_init(void) {
+  common_peripherals_init();
+  peripherals.common_peripherals = get_common_peripherals();
+
+  adc1_init();
+  i2c1_init();
+  spi3_init();
+  dma_init();
+  gpio_init();
+}
 
 void adc1_init(void) {
   ADC_HandleTypeDef* hadc1 = &peripherals.hadc1;
@@ -85,25 +102,6 @@ void adc1_init(void) {
   }
 }
 
-void can_init(void) {
-  CAN_HandleTypeDef* hcan = &peripherals.hcan;
-  hcan->Instance = CAN;
-  hcan->Init.Prescaler = 18;  // NOLINT
-  hcan->Init.Mode = CAN_MODE_NORMAL;
-  hcan->Init.SyncJumpWidth = CAN_SJW_1TQ;
-  hcan->Init.TimeSeg1 = CAN_BS1_13TQ;
-  hcan->Init.TimeSeg2 = CAN_BS2_2TQ;
-  hcan->Init.TimeTriggeredMode = DISABLE;
-  hcan->Init.AutoBusOff = DISABLE;
-  hcan->Init.AutoWakeUp = DISABLE;
-  hcan->Init.AutoRetransmission = DISABLE;
-  hcan->Init.ReceiveFifoLocked = DISABLE;
-  hcan->Init.TransmitFifoPriority = DISABLE;
-  if (HAL_CAN_Init(hcan) != HAL_OK) {
-    error();
-  }
-}
-
 void i2c1_init(void) {
   I2C_HandleTypeDef* hi2c1 = &peripherals.hi2c1;
   hi2c1->Instance = I2C1;
@@ -132,28 +130,6 @@ void i2c1_init(void) {
   }
 }
 
-void spi1_init(void) {
-  SPI_HandleTypeDef* hspi1 = &peripherals.hspi1;
-  /* SPI1 parameter configuration*/
-  hspi1->Instance = SPI1;
-  hspi1->Init.Mode = SPI_MODE_MASTER;
-  hspi1->Init.Direction = SPI_DIRECTION_2LINES;
-  hspi1->Init.DataSize = SPI_DATASIZE_4BIT;
-  hspi1->Init.CLKPolarity = SPI_POLARITY_LOW;
-  hspi1->Init.CLKPhase = SPI_PHASE_1EDGE;
-  hspi1->Init.NSS = SPI_NSS_HARD_OUTPUT;
-  hspi1->Init.BaudRatePrescaler = SPI_BAUDRATEPRESCALER_4;
-  hspi1->Init.FirstBit = SPI_FIRSTBIT_MSB;
-  hspi1->Init.TIMode = SPI_TIMODE_DISABLE;
-  hspi1->Init.CRCCalculation = SPI_CRCCALCULATION_DISABLE;
-  hspi1->Init.CRCPolynomial = 7;  // NOLINT
-  hspi1->Init.CRCLength = SPI_CRC_LENGTH_DATASIZE;
-  hspi1->Init.NSSPMode = SPI_NSS_PULSE_ENABLE;
-  if (HAL_SPI_Init(hspi1) != HAL_OK) {
-    error();
-  }
-}
-
 void spi3_init(void) {
   SPI_HandleTypeDef* hspi3 = &peripherals.hspi3;
   /* SPI3 parameter configuration*/
@@ -172,23 +148,6 @@ void spi3_init(void) {
   hspi3->Init.CRCLength = SPI_CRC_LENGTH_DATASIZE;
   hspi3->Init.NSSPMode = SPI_NSS_PULSE_ENABLE;
   if (HAL_SPI_Init(hspi3) != HAL_OK) {
-    error();
-  }
-}
-
-void uart1_init(void) {
-  UART_HandleTypeDef* huart1 = &peripherals.huart1;
-  huart1->Instance = USART1;
-  huart1->Init.BaudRate = 115200;  // NOLINT
-  huart1->Init.WordLength = UART_WORDLENGTH_8B;
-  huart1->Init.StopBits = UART_STOPBITS_1;
-  huart1->Init.Parity = UART_PARITY_NONE;
-  huart1->Init.Mode = UART_MODE_TX_RX;
-  huart1->Init.HwFlowCtl = UART_HWCONTROL_NONE;
-  huart1->Init.OverSampling = UART_OVERSAMPLING_16;
-  huart1->Init.OneBitSampling = UART_ONE_BIT_SAMPLE_DISABLE;
-  huart1->AdvancedInit.AdvFeatureInit = UART_ADVFEATURE_NO_INIT;
-  if (HAL_UART_Init(huart1) != HAL_OK) {
     error();
   }
 }
@@ -227,14 +186,6 @@ void gpio_init(void) {
   /*Configure GPIO pin Output Level */
   HAL_GPIO_WritePin(SPI3_NSS_GPIO_Port, SPI3_NSS_Pin, GPIO_PIN_RESET);
 
-  /*Configure GPIO pins : SWITCH4_PIN2_Pin SWITCH2_PIN1_Pin SWITCH2_PIN2_Pin
-     CAN_FD_INT_Pin CAN_FD_SOF_Pin */
-  GPIO_InitStruct.Pin = SWITCH4_PIN2_Pin | SWITCH2_PIN1_Pin | SWITCH2_PIN2_Pin |
-                        CAN_FD_INT_Pin | CAN_FD_SOF_Pin;
-  GPIO_InitStruct.Mode = GPIO_MODE_INPUT;
-  GPIO_InitStruct.Pull = GPIO_NOPULL;
-  HAL_GPIO_Init(GPIOC, &GPIO_InitStruct);
-
   /*Configure GPIO pins : GPIO0_Pin GPIO1_Pin GPIO2_Pin GPIO3_Pin
                            GPIO_PWRON_1_Pin GPIO_PWRON_2_Pin */
   GPIO_InitStruct.Pin = GPIO0_Pin | GPIO1_Pin | GPIO2_Pin | GPIO3_Pin |
@@ -243,14 +194,6 @@ void gpio_init(void) {
   GPIO_InitStruct.Pull = GPIO_NOPULL;
   GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_LOW;
   HAL_GPIO_Init(GPIOC, &GPIO_InitStruct);
-
-  /*Configure GPIO pins : SWITCH3_PIN1_Pin SWITCH3_PIN2_Pin SWITCH4_PIN1_Pin
-     CAN_FD_INT1_Pin CAN_FD_INT0_Pin */
-  GPIO_InitStruct.Pin = SWITCH3_PIN1_Pin | SWITCH3_PIN2_Pin | SWITCH4_PIN1_Pin |
-                        CAN_FD_INT1_Pin | CAN_FD_INT0_Pin;
-  GPIO_InitStruct.Mode = GPIO_MODE_INPUT;
-  GPIO_InitStruct.Pull = GPIO_NOPULL;
-  HAL_GPIO_Init(GPIOA, &GPIO_InitStruct);
 
   /*Configure GPIO pins : GPIO_PWRON_3_Pin GPIO_PWRON_4_Pin VDD_IO_LEVEL_Pin */
   GPIO_InitStruct.Pin = GPIO_PWRON_3_Pin | GPIO_PWRON_4_Pin | VDD_IO_LEVEL_Pin;
@@ -358,22 +301,8 @@ void HAL_ADC_MspDeInit(ADC_HandleTypeDef* hadc) {
  * @retval None
  */
 void HAL_CAN_MspInit(CAN_HandleTypeDef* hcan) {
-  GPIO_InitTypeDef GPIO_InitStruct;
   if (hcan->Instance == CAN) {
-    /* Peripheral clock enable */
-    __HAL_RCC_CAN1_CLK_ENABLE();
-
-    __HAL_RCC_GPIOB_CLK_ENABLE();
-    /**CAN GPIO Configuration
-    PB8     ------> CAN_RX
-    PB9     ------> CAN_TX
-    */
-    GPIO_InitStruct.Pin = GPIO_PIN_8 | GPIO_PIN_9;
-    GPIO_InitStruct.Mode = GPIO_MODE_AF_PP;
-    GPIO_InitStruct.Pull = GPIO_NOPULL;
-    GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_HIGH;
-    GPIO_InitStruct.Alternate = GPIO_AF9_CAN;
-    HAL_GPIO_Init(GPIOB, &GPIO_InitStruct);
+    can_msp_init();
   }
 }
 
@@ -385,14 +314,7 @@ void HAL_CAN_MspInit(CAN_HandleTypeDef* hcan) {
  */
 void HAL_CAN_MspDeInit(CAN_HandleTypeDef* hcan) {
   if (hcan->Instance == CAN) {
-    /* Peripheral clock disable */
-    __HAL_RCC_CAN1_CLK_DISABLE();
-
-    /**CAN GPIO Configuration
-    PB8     ------> CAN_RX
-    PB9     ------> CAN_TX
-    */
-    HAL_GPIO_DeInit(GPIOB, GPIO_PIN_8 | GPIO_PIN_9);
+    can_msp_deinit();
   }
 }
 
@@ -452,31 +374,7 @@ void HAL_I2C_MspDeInit(I2C_HandleTypeDef* hi2c) {
 void HAL_SPI_MspInit(SPI_HandleTypeDef* hspi) {
   GPIO_InitTypeDef GPIO_InitStruct;
   if (hspi->Instance == SPI1) {
-    /* Peripheral clock enable */
-    __HAL_RCC_SPI1_CLK_ENABLE();
-
-    __HAL_RCC_GPIOA_CLK_ENABLE();
-    __HAL_RCC_GPIOB_CLK_ENABLE();
-    /**SPI1 GPIO Configuration
-    PA15     ------> SPI1_NSS
-    PB3     ------> SPI1_SCK
-    PB4     ------> SPI1_MISO
-    PB5     ------> SPI1_MOSI
-    */
-    GPIO_InitStruct.Pin = CAN_FD_SPI_CS_Pin;
-    GPIO_InitStruct.Mode = GPIO_MODE_AF_PP;
-    GPIO_InitStruct.Pull = GPIO_NOPULL;
-    GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_HIGH;
-    GPIO_InitStruct.Alternate = GPIO_AF5_SPI1;
-    HAL_GPIO_Init(CAN_FD_SPI_CS_GPIO_Port, &GPIO_InitStruct);
-
-    GPIO_InitStruct.Pin =
-        CAN_FD_SPI_SCK_Pin | CAN_FD_SPI_MISO_Pin | CAN_FD_SPI_MOSI_Pin;
-    GPIO_InitStruct.Mode = GPIO_MODE_AF_PP;
-    GPIO_InitStruct.Pull = GPIO_NOPULL;
-    GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_HIGH;
-    GPIO_InitStruct.Alternate = GPIO_AF5_SPI1;
-    HAL_GPIO_Init(GPIOB, &GPIO_InitStruct);
+    spi1_msp_init();
 
   } else if (hspi->Instance == SPI3) {
     /* Peripheral clock enable */
@@ -505,19 +403,7 @@ void HAL_SPI_MspInit(SPI_HandleTypeDef* hspi) {
  */
 void HAL_SPI_MspDeInit(SPI_HandleTypeDef* hspi) {
   if (hspi->Instance == SPI1) {
-    /* Peripheral clock disable */
-    __HAL_RCC_SPI1_CLK_DISABLE();
-
-    /**SPI1 GPIO Configuration
-    PA15     ------> SPI1_NSS
-    PB3     ------> SPI1_SCK
-    PB4     ------> SPI1_MISO
-    PB5     ------> SPI1_MOSI
-    */
-    HAL_GPIO_DeInit(CAN_FD_SPI_CS_GPIO_Port, CAN_FD_SPI_CS_Pin);
-
-    HAL_GPIO_DeInit(
-        GPIOB, CAN_FD_SPI_SCK_Pin | CAN_FD_SPI_MISO_Pin | CAN_FD_SPI_MOSI_Pin);
+    spi1_msp_deinit();
 
   } else if (hspi->Instance == SPI3) {
     /* Peripheral clock disable */
@@ -539,26 +425,8 @@ void HAL_SPI_MspDeInit(SPI_HandleTypeDef* hspi) {
  * @retval None
  */
 void HAL_UART_MspInit(UART_HandleTypeDef* huart) {
-  GPIO_InitTypeDef GPIO_InitStruct;
   if (huart->Instance == USART1) {
-    /* Peripheral clock enable */
-    __HAL_RCC_USART1_CLK_ENABLE();
-
-    __HAL_RCC_GPIOA_CLK_ENABLE();
-    /**USART1 GPIO Configuration
-    PA9     ------> USART1_TX
-    PA10     ------> USART1_RX
-    */
-    GPIO_InitStruct.Pin = GPIO_PIN_9 | GPIO_PIN_10;
-    GPIO_InitStruct.Mode = GPIO_MODE_AF_PP;
-    GPIO_InitStruct.Pull = GPIO_NOPULL;
-    GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_HIGH;
-    GPIO_InitStruct.Alternate = GPIO_AF7_USART1;
-    HAL_GPIO_Init(GPIOA, &GPIO_InitStruct);
-
-    /* USART1 interrupt Init */
-    HAL_NVIC_SetPriority(USART1_IRQn, USART1_IRQ_PRIORITY, 0);
-    HAL_NVIC_EnableIRQ(USART1_IRQn);
+    uart1_msp_init();
   }
 }
 
@@ -570,16 +438,6 @@ void HAL_UART_MspInit(UART_HandleTypeDef* huart) {
  */
 void HAL_UART_MspDeInit(UART_HandleTypeDef* huart) {
   if (huart->Instance == USART1) {
-    /* Peripheral clock disable */
-    __HAL_RCC_USART1_CLK_DISABLE();
-
-    /**USART1 GPIO Configuration
-    PA9     ------> USART1_TX
-    PA10     ------> USART1_RX
-    */
-    HAL_GPIO_DeInit(GPIOA, GPIO_PIN_9 | GPIO_PIN_10);
-
-    /* USART1 interrupt DeInit */
-    HAL_NVIC_DisableIRQ(USART1_IRQn);
+    uart1_msp_deinit();
   }
 }
