+++ apps/servo-node/include/interrupts.h
@@ -14,7 +14,6 @@ extern "C" {
 
 void DMA1_Channel1_IRQHandler(void);
 void DMA2_Channel1_IRQHandler(void);
-void USB_HP_CAN_TX_IRQHandler(void);
 
 #ifdef __cplusplus
 }
