+++ apps/servo-node/src/peripherals.c
@@ -13,7 +13,30 @@ static peripherals_t peripherals;
 
 static uint32_t HAL_RCC_ADC12_CLK_ENABLED = 0;
 
+void gpio_init(void);
+void dma_init(void);
+void adc1_init(void);
+void adc2_init(void);
+void i2c1_init(void);
+void i2c3_init(void);
+void spi3_init(void);
+void tim1_init(void);
+
 peripherals_t* get_peripherals(void) { return &peripherals; }
+
+void peripherals_init(void) {
+  common_peripherals_init();
+  peripherals.common_peripherals = get_common_peripherals();
+
+  gpio_init();
+  dma_init();
+  adc1_init();
+  adc2_init();
+  i2c1_init();
+  i2c3_init();
+  spi3_init();
+  tim1_init();
+}
 
 void HAL_TIM_MspPostInit(TIM_HandleTypeDef* htim);
 
@@ -131,25 +154,6 @@ void adc2_init(void) {
   }
 }
 
-void can_init(void) {
-  CAN_HandleTypeDef* hcan = &peripherals.hcan;
-  hcan->Instance = CAN;
-  hcan->Init.Prescaler = 18;  // NOLINT
-  hcan->Init.Mode = CAN_MODE_NORMAL;
-  hcan->Init.SyncJumpWidth = CAN_SJW_1TQ;
-  hcan->Init.TimeSeg1 = CAN_BS1_13TQ;
-  hcan->Init.TimeSeg2 = CAN_BS2_2TQ;
-  hcan->Init.TimeTriggeredMode = DISABLE;
-  hcan->Init.AutoBusOff = DISABLE;
-  hcan->Init.AutoWakeUp = DISABLE;
-  hcan->Init.AutoRetransmission = DISABLE;
-  hcan->Init.ReceiveFifoLocked = DISABLE;
-  hcan->Init.TransmitFifoPriority = DISABLE;
-  if (HAL_CAN_Init(hcan) != HAL_OK) {
-    error();
-  }
-}
-
 void i2c1_init(void) {
   I2C_HandleTypeDef* hi2c1 = &peripherals.hi2c1;
   hi2c1->Instance = I2C1;
@@ -202,27 +206,6 @@ void i2c3_init(void) {
   /** Configure Digital filter
    */
   if (HAL_I2CEx_ConfigDigitalFilter(hi2c3, 0) != HAL_OK) {
-    error();
-  }
-}
-
-void spi1_init(void) {
-  SPI_HandleTypeDef* hspi1 = &peripherals.hspi1;
-  hspi1->Instance = SPI1;
-  hspi1->Init.Mode = SPI_MODE_MASTER;
-  hspi1->Init.Direction = SPI_DIRECTION_2LINES;
-  hspi1->Init.DataSize = SPI_DATASIZE_4BIT;
-  hspi1->Init.CLKPolarity = SPI_POLARITY_LOW;
-  hspi1->Init.CLKPhase = SPI_PHASE_1EDGE;
-  hspi1->Init.NSS = SPI_NSS_HARD_OUTPUT;
-  hspi1->Init.BaudRatePrescaler = SPI_BAUDRATEPRESCALER_4;
-  hspi1->Init.FirstBit = SPI_FIRSTBIT_MSB;
-  hspi1->Init.TIMode = SPI_TIMODE_DISABLE;
-  hspi1->Init.CRCCalculation = SPI_CRCCALCULATION_DISABLE;
-  hspi1->Init.CRCPolynomial = 7;  // NOLINT
-  hspi1->Init.CRCLength = SPI_CRC_LENGTH_DATASIZE;
-  hspi1->Init.NSSPMode = SPI_NSS_PULSE_ENABLE;
-  if (HAL_SPI_Init(hspi1) != HAL_OK) {
     error();
   }
 }
@@ -305,23 +288,6 @@ void tim1_init(void) {
   HAL_TIM_MspPostInit(htim1);
 }
 
-void uart1_init(void) {
-  UART_HandleTypeDef* huart1 = &peripherals.huart1;
-  huart1->Instance = USART1;
-  huart1->Init.BaudRate = 115200;  // NOLINT
-  huart1->Init.WordLength = UART_WORDLENGTH_8B;
-  huart1->Init.StopBits = UART_STOPBITS_1;
-  huart1->Init.Parity = UART_PARITY_NONE;
-  huart1->Init.Mode = UART_MODE_TX_RX;
-  huart1->Init.HwFlowCtl = UART_HWCONTROL_NONE;
-  huart1->Init.OverSampling = UART_OVERSAMPLING_16;
-  huart1->Init.OneBitSampling = UART_ONE_BIT_SAMPLE_DISABLE;
-  huart1->AdvancedInit.AdvFeatureInit = UART_ADVFEATURE_NO_INIT;
-  if (HAL_UART_Init(huart1) != HAL_OK) {
-    error();
-  }
-}
-
 void dma_init(void) {
   /* DMA controller clock enable */
   __HAL_RCC_DMA2_CLK_ENABLE();
@@ -390,18 +356,6 @@ void gpio_init(void) {
   GPIO_InitStruct.Mode = GPIO_MODE_INPUT;
   GPIO_InitStruct.Pull = GPIO_NOPULL;
   HAL_GPIO_Init(H_BRIDGE_nFAULT_GPIO_Port, &GPIO_InitStruct);
-
-  /*Configure GPIO pins : CAN_FD_INT_Pin CAN_FD_SOF_Pin */
-  GPIO_InitStruct.Pin = CAN_FD_INT_Pin | CAN_FD_SOF_Pin;
-  GPIO_InitStruct.Mode = GPIO_MODE_INPUT;
-  GPIO_InitStruct.Pull = GPIO_NOPULL;
-  HAL_GPIO_Init(GPIOC, &GPIO_InitStruct);
-
-  /*Configure GPIO pins : CAN_FD_INT1_Pin CAN_FD_INT0_Pin */
-  GPIO_InitStruct.Pin = CAN_FD_INT1_Pin | CAN_FD_INT0_Pin;
-  GPIO_InitStruct.Mode = GPIO_MODE_INPUT;
-  GPIO_InitStruct.Pull = GPIO_NOPULL;
-  HAL_GPIO_Init(GPIOA, &GPIO_InitStruct);
 
   /*Configure GPIO pin : SPI3_NSS_Pin */
   GPIO_InitStruct.Pin = SPI3_NSS_Pin;
@@ -556,26 +510,8 @@ void HAL_ADC_MspDeInit(ADC_HandleTypeDef* hadc) {
  * @retval None
  */
 void HAL_CAN_MspInit(CAN_HandleTypeDef* hcan) {
-  GPIO_InitTypeDef GPIO_InitStruct;
   if (hcan->Instance == CAN) {
-    /* Peripheral clock enable */
-    __HAL_RCC_CAN1_CLK_ENABLE();
-
-    __HAL_RCC_GPIOB_CLK_ENABLE();
-    /**CAN GPIO Configuration
-    PB8     ------> CAN_RX
-    PB9     ------> CAN_TX
-    */
-    GPIO_InitStruct.Pin = GPIO_PIN_8 | GPIO_PIN_9;
-    GPIO_InitStruct.Mode = GPIO_MODE_AF_PP;
-    GPIO_InitStruct.Pull = GPIO_NOPULL;
-    GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_HIGH;
-    GPIO_InitStruct.Alternate = GPIO_AF9_CAN;
-    HAL_GPIO_Init(GPIOB, &GPIO_InitStruct);
-
-    /* CAN interrupt Init */
-    HAL_NVIC_SetPriority(USB_HP_CAN_TX_IRQn, USB_HP_CAN_TX_IRQ_PRIORITY, 0);
-    HAL_NVIC_EnableIRQ(USB_HP_CAN_TX_IRQn);
+    can_msp_init();
   }
 }
 
@@ -587,17 +523,7 @@ void HAL_CAN_MspInit(CAN_HandleTypeDef* hcan) {
  */
 void HAL_CAN_MspDeInit(CAN_HandleTypeDef* hcan) {
   if (hcan->Instance == CAN) {
-    /* Peripheral clock disable */
-    __HAL_RCC_CAN1_CLK_DISABLE();
-
-    /**CAN GPIO Configuration
-    PB8     ------> CAN_RX
-    PB9     ------> CAN_TX
-    */
-    HAL_GPIO_DeInit(GPIOB, GPIO_PIN_8 | GPIO_PIN_9);
-
-    /* CAN interrupt DeInit */
-    HAL_NVIC_DisableIRQ(USB_HP_CAN_TX_IRQn);
+    can_msp_deinit();
   }
 }
 
@@ -693,31 +619,7 @@ void HAL_I2C_MspDeInit(I2C_HandleTypeDef* hi2c) {
 void HAL_SPI_MspInit(SPI_HandleTypeDef* hspi) {
   GPIO_InitTypeDef GPIO_InitStruct;
   if (hspi->Instance == SPI1) {
-    /* Peripheral clock enable */
-    __HAL_RCC_SPI1_CLK_ENABLE();
-
-    __HAL_RCC_GPIOA_CLK_ENABLE();
-    __HAL_RCC_GPIOB_CLK_ENABLE();
-    /**SPI1 GPIO Configuration
-    PA15     ------> SPI1_NSS
-    PB3     ------> SPI1_SCK
-    PB4     ------> SPI1_MISO
-    PB5     ------> SPI1_MOSI
-    */
-    GPIO_InitStruct.Pin = CAN_FD_SPI_CS_Pin;
-    GPIO_InitStruct.Mode = GPIO_MODE_AF_PP;
-    GPIO_InitStruct.Pull = GPIO_NOPULL;
-    GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_HIGH;
-    GPIO_InitStruct.Alternate = GPIO_AF5_SPI1;
-    HAL_GPIO_Init(CAN_FD_SPI_CS_GPIO_Port, &GPIO_InitStruct);
-
-    GPIO_InitStruct.Pin =
-        CAN_FD_SPI_SCK_Pin | CAN_FD_SPI_MISO_Pin | CAN_FD_SPI_MOSI_Pin;
-    GPIO_InitStruct.Mode = GPIO_MODE_AF_PP;
-    GPIO_InitStruct.Pull = GPIO_NOPULL;
-    GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_HIGH;
-    GPIO_InitStruct.Alternate = GPIO_AF5_SPI1;
-    HAL_GPIO_Init(GPIOB, &GPIO_InitStruct);
+    spi1_msp_init();
 
   } else if (hspi->Instance == SPI3) {
     /* Peripheral clock enable */
@@ -746,19 +648,7 @@ void HAL_SPI_MspInit(SPI_HandleTypeDef* hspi) {
  */
 void HAL_SPI_MspDeInit(SPI_HandleTypeDef* hspi) {
   if (hspi->Instance == SPI1) {
-    /* Peripheral clock disable */
-    __HAL_RCC_SPI1_CLK_DISABLE();
-
-    /**SPI1 GPIO Configuration
-    PA15     ------> SPI1_NSS
-    PB3     ------> SPI1_SCK
-    PB4     ------> SPI1_MISO
-    PB5     ------> SPI1_MOSI
-    */
-    HAL_GPIO_DeInit(CAN_FD_SPI_CS_GPIO_Port, CAN_FD_SPI_CS_Pin);
-
-    HAL_GPIO_DeInit(
-        GPIOB, CAN_FD_SPI_SCK_Pin | CAN_FD_SPI_MISO_Pin | CAN_FD_SPI_MOSI_Pin);
+    spi1_msp_deinit();
 
   } else if (hspi->Instance == SPI3) {
     /* Peripheral clock disable */
@@ -821,26 +711,8 @@ void HAL_TIM_Base_MspDeInit(TIM_HandleTypeDef* htim_base) {
  * @retval None
  */
 void HAL_UART_MspInit(UART_HandleTypeDef* huart) {
-  GPIO_InitTypeDef GPIO_InitStruct;
   if (huart->Instance == USART1) {
-    /* Peripheral clock enable */
-    __HAL_RCC_USART1_CLK_ENABLE();
-
-    __HAL_RCC_GPIOA_CLK_ENABLE();
-    /**USART1 GPIO Configuration
-    PA9     ------> USART1_TX
-    PA10     ------> USART1_RX
-    */
-    GPIO_InitStruct.Pin = GPIO_PIN_9 | GPIO_PIN_10;
-    GPIO_InitStruct.Mode = GPIO_MODE_AF_PP;
-    GPIO_InitStruct.Pull = GPIO_NOPULL;
-    GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_HIGH;
-    GPIO_InitStruct.Alternate = GPIO_AF7_USART1;
-    HAL_GPIO_Init(GPIOA, &GPIO_InitStruct);
-
-    /* USART1 interrupt Init */
-    HAL_NVIC_SetPriority(USART1_IRQn, USART1_IRQ_PRIORITY, 0);
-    HAL_NVIC_EnableIRQ(USART1_IRQn);
+    uart1_msp_init();
   }
 }
 
@@ -852,16 +724,6 @@ void HAL_UART_MspInit(UART_HandleTypeDef* huart) {
  */
 void HAL_UART_MspDeInit(UART_HandleTypeDef* huart) {
   if (huart->Instance == USART1) {
-    /* Peripheral clock disable */
-    __HAL_RCC_USART1_CLK_DISABLE();
-
-    /**USART1 GPIO Configuration
-    PA9     ------> USART1_TX
-    PA10     ------> USART1_RX
-    */
-    HAL_GPIO_DeInit(GPIOA, GPIO_PIN_9 | GPIO_PIN_10);
-
-    /* USART1 interrupt DeInit */
-    HAL_NVIC_DisableIRQ(USART1_IRQn);
+    uart1_msp_deinit();
   }
 }
