+++ libs/stm32-common/include/common-interrupts.h
@@ -13,6 +13,9 @@ void UsageFault_Handler(void);
 void DebugMon_Handler(void);
 void SysTick_Handler(void);
 
+void USART1_IRQHandler(void);
+void USB_LP_CAN_RX0_IRQHandler(void);
+
 #ifdef __cplusplus
 }
 #endif
