+++ libs/stm32-common/include/common-ports.h
@@ -1,0 +1,31 @@
+#ifndef COMMON_PORTS_H
+#define COMMON_PORTS_H
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#include "stm32f3xx_hal.h"
+
+#define CAN_FD_INT_PIN GPIO_PIN_7
+#define CAN_FD_INT_GPIO_PORT GPIOC
+#define CAN_FD_SOF_PIN GPIO_PIN_8
+#define CAN_FD_SOF_GPIO_PORT GPIOC
+#define CAN_FD_INT1_PIN GPIO_PIN_11
+#define CAN_FD_INT1_GPIO_PORT GPIOA
+#define CAN_FD_INT0_PIN GPIO_PIN_12
+#define CAN_FD_INT0_GPIO_PORT GPIOA
+#define CAN_FD_SPI_CS_PIN GPIO_PIN_15
+#define CAN_FD_SPI_CS_GPIO_PORT GPIOA
+#define CAN_FD_SPI_SCK_PIN GPIO_PIN_3
+#define CAN_FD_SPI_SCK_GPIO_PORT GPIOB
+#define CAN_FD_SPI_MISO_PIN GPIO_PIN_4
+#define CAN_FD_SPI_MISO_GPIO_PORT GPIOB
+#define CAN_FD_SPI_MOSI_PIN GPIO_PIN_5
+#define CAN_FD_SPI_MOSI_GPIO_PORT GPIOB
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* COMMON_PORTS_H */
