+++ libs/stm32-common/meson.build
@@ -4,6 +4,7 @@ stm32_common_src = files(
   'src' / 'assert.c',
   'src' / 'clock.c',
   'src' / 'common-interrupts.c',
+  'src' / 'common-peripherals.c',
   'src' / 'error.c',
   'src' / 'flash.c',
   'src' / 'freertos.c',
@@ -22,6 +23,7 @@ stm32_common_tidy_files = files(
   'src' / 'assert.c',
   'src' / 'clock.c',
   'src' / 'common-interrupts.c',
+  'src' / 'common-peripherals.c',
   'src' / 'error.c',
   'src' / 'flash.c',
   'src' / 'freertos.c',
